// src/app/forms/appointment/user.interface.ts
export interface User {
    strName: string;
    strPhone: string;
    strEmail: string;
    strContactMethod: string;
    
    strMessage: string;

    isMonday: boolean;
    isTuesday: boolean;
    isWednesday: boolean;
    isThursday: boolean;
    isFriday: boolean;
    strTimeOfDay: string;

    googleResponse: string;
}